Pluralize strategy mutations
