+++ packages/strategy-email/src/interface/EmailAuthority.tsx
