+++ packages/strategy-email/src/interface/index.ts
