+++ packages/strategy-email/src/server/graphql/mutation/GraphQLCreateEmailCredentialInput.ts
@@ -1,0 +1,34 @@
+import {
+  GraphQLID,
+  GraphQLNonNull,
+  GraphQLString,
+  GraphQLBoolean,
+  GraphQLInputObjectType
+} from "graphql";
+
+export const GraphQLCreateEmailCredentialInput = new GraphQLInputObjectType({
+  name: "CreateEmailCredentialInput",
+  fields: () => ({
+    enabled: {
+      type: new GraphQLNonNull(GraphQLBoolean)
+    },
+    userId: {
+      type: new GraphQLNonNull(GraphQLID),
+      description: "The ID of the AuthX user who will own this credential."
+    },
+    authorityId: {
+      type: new GraphQLNonNull(GraphQLID),
+      description:
+        "The ID of the AuthX email authority that can verify this credential."
+    },
+    email: {
+      type: new GraphQLNonNull(GraphQLID),
+      description: "The email address of the user."
+    },
+    proof: {
+      type: GraphQLString,
+      description:
+        "This is a unique code that was sent by the authority to prove control of the email address."
+    }
+  })
+});
