+++ packages/strategy-email/src/server/graphql/mutation/updateEmailAuthority.ts
@@ -1,13 +1,5 @@
 import v4 from "uuid/v4";
-import {
-  GraphQLBoolean,
-  GraphQLFieldConfig,
-  GraphQLID,
-  GraphQLNonNull,
-  GraphQLList,
-  GraphQLInt,
-  GraphQLString
-} from "graphql";
+import { GraphQLFieldConfig, GraphQLNonNull, GraphQLList } from "graphql";
 
 import {
   Context,
@@ -17,95 +9,40 @@ import {
 } from "@authx/authx";
 import { EmailAuthority } from "../../model";
 import { GraphQLEmailAuthority } from "../GraphQLEmailAuthority";
+import { GraphQLUpdateEmailAuthorityInput } from "./GraphQLUpdateEmailAuthorityInput";
 
 export const updateEmailAuthority: GraphQLFieldConfig<
   any,
   {
-    id: string;
-    enabled: null | boolean;
-    name: null | string;
-    description: null | string;
-    privateKey: null | string;
-    addPublicKeys: null | string[];
-    removePublicKeys: null | string[];
-    proofValidityDuration: null | number;
-    authenticationEmailSubject: null | string;
-    authenticationEmailText: null | string;
-    authenticationEmailHtml: null | string;
-    verificationEmailSubject: null | string;
-    verificationEmailText: null | string;
-    verificationEmailHtml: null | string;
+    authorities: {
+      id: string;
+      enabled: null | boolean;
+      name: null | string;
+      description: null | string;
+      privateKey: null | string;
+      addPublicKeys: null | string[];
+      removePublicKeys: null | string[];
+      proofValidityDuration: null | number;
+      authenticationEmailSubject: null | string;
+      authenticationEmailText: null | string;
+      authenticationEmailHtml: null | string;
+      verificationEmailSubject: null | string;
+      verificationEmailText: null | string;
+      verificationEmailHtml: null | string;
+    }[];
   },
   Context
 > = {
   type: GraphQLEmailAuthority,
   description: "Update a new authority.",
   args: {
-    id: {
-      type: new GraphQLNonNull(GraphQLID)
-    },
-    enabled: {
-      type: GraphQLBoolean
-    },
-    name: {
-      type: GraphQLString,
-      description: "The name of the authority."
-    },
-    description: {
-      type: GraphQLString,
-      description: "A description of the authority."
-    },
-    privateKey: {
-      type: GraphQLString,
-      description:
-        "The RS512 private key that will be used to sign the proofs sent to verify ownership of email addresses."
-    },
-    addPublicKeys: {
-      type: new GraphQLList(new GraphQLNonNull(GraphQLString)),
-      description:
-        "Add a key to the list of RS512 public keys that will be used to verify the proofs sent to verify ownership of email addresses. This allows private keys to be rotated without invalidating current proofs."
-    },
-    removePublicKeys: {
-      type: new GraphQLList(new GraphQLNonNull(GraphQLString)),
-      description:
-        "Remove a key from the list of RS512 public keys that will be used to verify the proofs sent to verify ownership of email addresses."
-    },
-    proofValidityDuration: {
-      type: GraphQLInt,
-      description: "Time in seconds until the email link expires."
-    },
-    authenticationEmailSubject: {
-      type: GraphQLString,
-      description:
-        "Authentication Email Subject. Handlebars template used to generate the email subject. Provided `authorization`, `credential`, and `url`."
-    },
-    authenticationEmailText: {
-      type: GraphQLString,
-      description:
-        "Authentication Email Plain Text Body. Handlebars template used to generate the email plain text body. Provided `authorization`, `credential`, and `url`."
-    },
-    authenticationEmailHtml: {
-      type: GraphQLString,
-      description:
-        "Authentication Email HTML Body. Handlebars template used to generate the email HTML body. Provided `authorization`, `credential`, and `url`."
-    },
-    verificationEmailSubject: {
-      type: GraphQLString,
-      description:
-        "Verification Email Subject. Handlebars template used to generate the email subject. Provided `authorization`, `credential`, and `url`."
-    },
-    verificationEmailText: {
-      type: GraphQLString,
-      description:
-        "Verification Email Plain Text Body. Handlebars template used to generate the email plain text body. Provided `authorization`, `credential`, and `url`."
-    },
-    verificationEmailHtml: {
-      type: GraphQLString,
-      description:
-        "Verification Email HTML Body. Handlebars template used to generate the email HTML body. Provided `authorization`, `credential`, and `url`."
+    authorities: {
+      type: new GraphQLNonNull(
+        new GraphQLList(new GraphQLNonNull(GraphQLUpdateEmailAuthorityInput))
+      )
     }
   },
-  async resolve(source, args, context): Promise<EmailAuthority> {
+  async resolve(source, args, context): Promise<Promise<EmailAuthority>[]> {
     const {
       pool,
       authorization: a,
@@ -119,115 +56,119 @@ export const updateEmailAuthority: GraphQLFieldConfig<
       );
     }
 
-    const tx = await pool.connect();
-    try {
-      await tx.query("BEGIN DEFERRABLE");
+    return args.authorities.map(async input => {
+      const tx = await pool.connect();
+      try {
+        await tx.query("BEGIN DEFERRABLE");
 
-      const before = await Authority.read(tx, args.id, authorityMap);
+        const before = await Authority.read(tx, input.id, authorityMap);
 
-      if (!(before instanceof EmailAuthority)) {
-        throw new NotFoundError(
-          "The authority uses a strategy other than email."
-        );
-      }
-
-      if (!(await before.isAccessibleBy(realm, a, tx, "write.basic"))) {
-        throw new ForbiddenError(
-          "You do not have permission to update this authority."
-        );
-      }
-
-      if (
-        (typeof args.privateKey === "string" ||
-          args.addPublicKeys ||
-          args.removePublicKeys ||
-          typeof args.proofValidityDuration === "number" ||
-          typeof args.authenticationEmailSubject === "string" ||
-          typeof args.authenticationEmailText === "string" ||
-          typeof args.authenticationEmailHtml === "string" ||
-          typeof args.verificationEmailSubject === "string" ||
-          typeof args.verificationEmailText === "string" ||
-          typeof args.verificationEmailHtml === "string") &&
-        !(await before.isAccessibleBy(realm, a, tx, "write.*"))
-      ) {
-        throw new ForbiddenError(
-          "You do not have permission to update this authority's details."
-        );
-      }
-
-      let publicKeys = [...before.details.publicKeys];
-
-      const { addPublicKeys } = args;
-      if (addPublicKeys) {
-        publicKeys = [...publicKeys, ...addPublicKeys];
-      }
-
-      const { removePublicKeys } = args;
-      if (removePublicKeys) {
-        publicKeys = publicKeys.filter(k => !removePublicKeys.includes(k));
-      }
-
-      const authority = await EmailAuthority.write(
-        tx,
-        {
-          ...before,
-          enabled:
-            typeof args.enabled === "boolean" ? args.enabled : before.enabled,
-          name: typeof args.name === "string" ? args.name : before.name,
-          description:
-            typeof args.description === "string"
-              ? args.description
-              : before.description,
-          details: {
-            privateKey:
-              typeof args.privateKey === "string"
-                ? args.privateKey
-                : before.details.privateKey,
-            publicKeys,
-            proofValidityDuration:
-              typeof args.proofValidityDuration === "number"
-                ? args.proofValidityDuration
-                : before.details.proofValidityDuration,
-            authenticationEmailSubject:
-              typeof args.authenticationEmailSubject === "string"
-                ? args.authenticationEmailSubject
-                : before.details.authenticationEmailSubject,
-            authenticationEmailText:
-              typeof args.authenticationEmailText === "string"
-                ? args.authenticationEmailText
-                : before.details.authenticationEmailText,
-            authenticationEmailHtml:
-              typeof args.authenticationEmailHtml === "string"
-                ? args.authenticationEmailHtml
-                : before.details.authenticationEmailHtml,
-            verificationEmailSubject:
-              typeof args.verificationEmailSubject === "string"
-                ? args.verificationEmailSubject
-                : before.details.verificationEmailSubject,
-            verificationEmailText:
-              typeof args.verificationEmailText === "string"
-                ? args.verificationEmailText
-                : before.details.verificationEmailText,
-            verificationEmailHtml:
-              typeof args.verificationEmailHtml === "string"
-                ? args.verificationEmailHtml
-                : before.details.verificationEmailHtml
-          }
-        },
-        {
-          recordId: v4(),
-          createdByAuthorizationId: a.id,
-          createdAt: new Date()
+        if (!(before instanceof EmailAuthority)) {
+          throw new NotFoundError(
+            "The authority uses a strategy other than email."
+          );
         }
-      );
 
-      await tx.query("COMMIT");
-      return authority;
-    } catch (error) {
-      await tx.query("ROLLBACK");
-      throw error;
-    } finally {
-      tx.release();
-    }
+        if (!(await before.isAccessibleBy(realm, a, tx, "write.basic"))) {
+          throw new ForbiddenError(
+            "You do not have permission to update this authority."
+          );
+        }
+
+        if (
+          (typeof input.privateKey === "string" ||
+            input.addPublicKeys ||
+            input.removePublicKeys ||
+            typeof input.proofValidityDuration === "number" ||
+            typeof input.authenticationEmailSubject === "string" ||
+            typeof input.authenticationEmailText === "string" ||
+            typeof input.authenticationEmailHtml === "string" ||
+            typeof input.verificationEmailSubject === "string" ||
+            typeof input.verificationEmailText === "string" ||
+            typeof input.verificationEmailHtml === "string") &&
+          !(await before.isAccessibleBy(realm, a, tx, "write.*"))
+        ) {
+          throw new ForbiddenError(
+            "You do not have permission to update this authority's details."
+          );
+        }
+
+        let publicKeys = [...before.details.publicKeys];
+
+        const { addPublicKeys } = args;
+        if (addPublicKeys) {
+          publicKeys = [...publicKeys, ...addPublicKeys];
+        }
+
+        const { removePublicKeys } = args;
+        if (removePublicKeys) {
+          publicKeys = publicKeys.filter(k => !removePublicKeys.includes(k));
+        }
+
+        const authority = await EmailAuthority.write(
+          tx,
+          {
+            ...before,
+            enabled:
+              typeof input.enabled === "boolean"
+                ? input.enabled
+                : before.enabled,
+            name: typeof input.name === "string" ? input.name : before.name,
+            description:
+              typeof input.description === "string"
+                ? input.description
+                : before.description,
+            details: {
+              privateKey:
+                typeof input.privateKey === "string"
+                  ? input.privateKey
+                  : before.details.privateKey,
+              publicKeys,
+              proofValidityDuration:
+                typeof input.proofValidityDuration === "number"
+                  ? input.proofValidityDuration
+                  : before.details.proofValidityDuration,
+              authenticationEmailSubject:
+                typeof input.authenticationEmailSubject === "string"
+                  ? input.authenticationEmailSubject
+                  : before.details.authenticationEmailSubject,
+              authenticationEmailText:
+                typeof input.authenticationEmailText === "string"
+                  ? input.authenticationEmailText
+                  : before.details.authenticationEmailText,
+              authenticationEmailHtml:
+                typeof input.authenticationEmailHtml === "string"
+                  ? input.authenticationEmailHtml
+                  : before.details.authenticationEmailHtml,
+              verificationEmailSubject:
+                typeof input.verificationEmailSubject === "string"
+                  ? input.verificationEmailSubject
+                  : before.details.verificationEmailSubject,
+              verificationEmailText:
+                typeof input.verificationEmailText === "string"
+                  ? input.verificationEmailText
+                  : before.details.verificationEmailText,
+              verificationEmailHtml:
+                typeof input.verificationEmailHtml === "string"
+                  ? input.verificationEmailHtml
+                  : before.details.verificationEmailHtml
+            }
+          },
+          {
+            recordId: v4(),
+            createdByAuthorizationId: a.id,
+            createdAt: new Date()
+          }
+        );
+
+        await tx.query("COMMIT");
+        return authority;
+      } catch (error) {
+        await tx.query("ROLLBACK");
+        throw error;
+      } finally {
+        tx.release();
+      }
+    });
   }
 };
