+++ packages/strategy-openid/src/interface/OpenIdAuthority.tsx
