+++ packages/strategy-openid/src/interface/index.ts
