+++ packages/strategy-openid/src/server/graphql/mutation/GraphQLCreateOpenIdAuthorityInput.ts
@@ -1,0 +1,73 @@
+import {
+  GraphQLID,
+  GraphQLNonNull,
+  GraphQLString,
+  GraphQLBoolean,
+  GraphQLInputObjectType,
+  GraphQLList
+} from "graphql";
+
+export const GraphQLCreateOpenIdAuthorityInput = new GraphQLInputObjectType({
+  name: "CreateOpenIdAuthorityInput",
+  fields: () => ({
+    enabled: {
+      type: GraphQLBoolean,
+      defaultValue: true
+    },
+    name: {
+      type: new GraphQLNonNull(GraphQLString),
+      description: "The name of the authority."
+    },
+    description: {
+      type: new GraphQLNonNull(GraphQLString),
+      description: "A description of the authority."
+    },
+    authUrl: {
+      type: new GraphQLNonNull(GraphQLString),
+      description: "The URL to which a user is directed to authenticate."
+    },
+    tokenUrl: {
+      type: new GraphQLNonNull(GraphQLString),
+      description:
+        "The URL used by AuthX to exchange an authorization code for an access token."
+    },
+    clientId: {
+      type: new GraphQLNonNull(GraphQLString),
+      description: "The client ID of AuthX in with OpenID provider."
+    },
+    clientSecret: {
+      type: new GraphQLNonNull(GraphQLString),
+      description: "The AuthX client secret with the OpenID provider."
+    },
+    restrictsAccountsToHostedDomains: {
+      type: new GraphQLNonNull(
+        new GraphQLList(new GraphQLNonNull(GraphQLString))
+      ) as any,
+      description: "Restrict to accounts controlled by these hosted domains.",
+      defaultValue: []
+    },
+    emailAuthorityId: {
+      type: GraphQLID,
+      description: "The ID of the email authority."
+    },
+    matchesUsersByEmail: {
+      type: GraphQLBoolean,
+      description:
+        "If no credential exists for the given OpenID provider, should we lookup the user by email address?",
+      defaultValue: false
+    },
+    createsUnmatchedUsers: {
+      type: GraphQLBoolean,
+      description:
+        "If no credential exists for the given OpenID provider, should we create a new one?",
+      defaultValue: false
+    },
+    assignsCreatedUsersToRoleIds: {
+      type: new GraphQLNonNull(
+        new GraphQLList(new GraphQLNonNull(GraphQLID))
+      ) as any,
+      description: "When a user is created, assign to these roles.",
+      defaultValue: []
+    }
+  })
+});
