--- packages/strategy-password/client.js
@@ -1,1 +1,0 @@
-module.exports = require("./dist/client");
