+++ packages/strategy-password/src/interface/PasswordAuthority.tsx
