+++ packages/strategy-password/src/interface/index.ts
