+++ packages/strategy-password/src/server/graphql/mutation/GraphQLUpdatePasswordCredentialInput.ts
@@ -1,0 +1,23 @@
+import {
+  GraphQLID,
+  GraphQLNonNull,
+  GraphQLString,
+  GraphQLBoolean,
+  GraphQLInputObjectType
+} from "graphql";
+
+export const GraphQLUpdatePasswordCredentialInput = new GraphQLInputObjectType({
+  name: "UpdatePasswordCredentialInput",
+  fields: () => ({
+    id: {
+      type: new GraphQLNonNull(GraphQLID)
+    },
+    enabled: {
+      type: GraphQLBoolean
+    },
+    password: {
+      type: GraphQLString,
+      description: "The plaintext password to use for this credential."
+    }
+  })
+});
