+++ packages/strategy-password/src/server/graphql/mutation/index.ts
@@ -1,5 +1,5 @@
 export * from "./authenticatePassword";
-export * from "./createPasswordAuthority";
-export * from "./createPasswordCredential";
-export * from "./updatePasswordAuthority";
-export * from "./updatePasswordCredential";
+export * from "./createPasswordAuthorities";
+export * from "./createPasswordCredentials";
+export * from "./updatePasswordAuthorities";
+export * from "./updatePasswordCredentials";
