--- packages/strategy-password/src/server/graphql/mutation/updatePasswordAuthority.ts
@@ -1,128 +1,0 @@
-import v4 from "uuid/v4";
-import {
-  GraphQLBoolean,
-  GraphQLFieldConfig,
-  GraphQLID,
-  GraphQLNonNull,
-  GraphQLInt,
-  GraphQLString
-} from "graphql";
-
-import {
-  Context,
-  Authority,
-  ForbiddenError,
-  NotFoundError
-} from "@authx/authx";
-import { PasswordAuthority } from "../../model";
-import { GraphQLPasswordAuthority } from "../GraphQLPasswordAuthority";
-
-export const updatePasswordAuthority: GraphQLFieldConfig<
-  any,
-  {
-    id: string;
-    enabled: null | boolean;
-    name: null | string;
-    description: null | string;
-    rounds: null | number;
-  },
-  Context
-> = {
-  type: GraphQLPasswordAuthority,
-  description: "Update a new authority.",
-  args: {
-    id: {
-      type: new GraphQLNonNull(GraphQLID)
-    },
-    enabled: {
-      type: GraphQLBoolean
-    },
-    name: {
-      type: GraphQLString,
-      description: "The name of the authority."
-    },
-    description: {
-      type: GraphQLString,
-      description: "The description of the authority."
-    },
-    rounds: {
-      type: GraphQLInt,
-      description:
-        "The number of bcrypt rounds to use when generating new hashes."
-    }
-  },
-  async resolve(source, args, context): Promise<PasswordAuthority> {
-    const {
-      pool,
-      authorization: a,
-      realm,
-      strategies: { authorityMap }
-    } = context;
-
-    if (!a) {
-      throw new ForbiddenError(
-        "You must be authenticated to update an authority."
-      );
-    }
-
-    const tx = await pool.connect();
-    try {
-      await tx.query("BEGIN DEFERRABLE");
-
-      const before = await Authority.read(tx, args.id, authorityMap);
-
-      if (!(before instanceof PasswordAuthority)) {
-        throw new NotFoundError("No password authority exists with this ID.");
-      }
-
-      if (!(await before.isAccessibleBy(realm, a, tx, "write.basic"))) {
-        throw new ForbiddenError(
-          "You do not have permission to update this authority."
-        );
-      }
-
-      if (
-        typeof args.rounds === "number" &&
-        !(await before.isAccessibleBy(realm, a, tx, "write.*"))
-      ) {
-        throw new ForbiddenError(
-          "You do not have permission to update this authority's details."
-        );
-      }
-
-      const authority = await PasswordAuthority.write(
-        tx,
-        {
-          ...before,
-          enabled:
-            typeof args.enabled === "boolean" ? args.enabled : before.enabled,
-          name: typeof args.name === "string" ? args.name : before.name,
-          description:
-            typeof args.description === "string"
-              ? args.description
-              : before.description,
-          details: {
-            ...before.details,
-            rounds:
-              typeof args.rounds === "number"
-                ? args.rounds
-                : before.details.rounds
-          }
-        },
-        {
-          recordId: v4(),
-          createdByAuthorizationId: a.id,
-          createdAt: new Date()
-        }
-      );
-
-      await tx.query("COMMIT");
-      return authority;
-    } catch (error) {
-      await tx.query("ROLLBACK");
-      throw error;
-    } finally {
-      tx.release();
-    }
-  }
-};
